arreglo de las versiones de hibernate

+++ src/main/java/com/innovaultmy/app/repositories/registros/InventarioFilamentoRepositorie.java
@@ -4,6 +4,6 @@ import com.innovaultmy.app.models.registros.InventarioFilamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface InventarioFilamentoRepositorie extends JpaRepository<InventarioFilamento, Integer> {
+public interface InventarioFilamentoRepositorie extends JpaRepository<InventarioFilamento, Long> {
 
 }
